Add files via upload
This program will allow the user to input from the keyboard
whether the last word to the following proverb should be party or country:
Now is the time for all good men to come to the aid of their ___"
Inputting a 1 will use the word party. Any other number will use the word country.